+++ relay-metrics/benches/aggregator.rs
@@ -6,8 +6,7 @@ use actix::prelude::*;
 use criterion::{criterion_group, criterion_main, BatchSize, BenchmarkId, Criterion};
 
 use relay_common::{ProjectKey, UnixTimestamp};
-use relay_metrics::{Aggregator, AggregatorConfig};
-use relay_metrics::{FlushBuckets, Metric, MetricValue};
+use relay_metrics::{Aggregator, AggregatorConfig, FlushBuckets, Metric, MetricValue};
 
 #[derive(Clone, Default)]
 struct TestReceiver;