+++ CHANGELOG.md
@@ -12,7 +12,7 @@
 - Generate a new profile ID when splitting a profile for multiple transactions. ([#1473](https://github.com/getsentry/relay/pull/1473))
 - Pin Rust version to 1.63.0 in Dockerfile. ([#1482](https://github.com/getsentry/relay/pull/1482))
 - Normalize measurement units in event payload. ([#1488](https://github.com/getsentry/relay/pull/1488))
-- Remove long-running futures from metrics flush ([#1492](https://github.com/getsentry/relay/pull/1492))
+- Remove long-running futures from metrics flush. ([#1492](https://github.com/getsentry/relay/pull/1492))
 
 ## 22.9.0